+++ CHANGELOG.md
@@ -10,6 +10,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 To see unreleased changes, please see the [CHANGELOG on the master branch](https://github.com/gufolabs/gufo_acme/blob/master/CHANGELOG.md) guide.
 
+## [Unreleased]
+
+## Fixed
+
+* Fixed `ACMEClient.from_state()` to return a proper subclass.
+* Fixed type annotation for `ACMEClient.__aenter__()` in subclasses.
+
 ## 0.1.0 - 2023-11-15
 
 * Initial release.